Complete CSS Excersises in 01-css-methods
Performed the required excersises using the CSS External, CSS Internal,
and CSS Inline methods. Created a new styles.css file to perform the
CSS External method on the <div> element. Added the styles element to
the <head> element to perform the CSS Internal method on the <p> element.
Added the style="" attribute to the <button> element to perform the CSS
Inline method on the element.